Move model to its own file
